Gating cleanup
